server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
ecdsa==0.19.0
executing==2.2.0
fastapi==0.115.8
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gotrue==2.11.3
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
ipyevents==2.0.2
ipympl==0.9.7
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mne==1.9.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2025.2
python-multipart==0.0.20
pyvista==0.44.1
pyvistaqt==0.11.2
PyYAML==6.0.2
QtPy==2.4.3
realtime==2.3.0
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scooby==0.10.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.6
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vtk==9.4.2
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xgboost==3.0.0
xvfbwrapper==0.2.10
yarl==1.18.3
